Update entity files.
Also fix a relation it depends on.
