+++ deepwell/src/models/alias.rs
@@ -21,18 +21,18 @@ pub struct Model {
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
-        belongs_to = "super::user::Entity",
+        belongs_to = "super::known_user::Entity",
         from = "Column::CreatedBy",
-        to = "super::user::Column::UserId",
+        to = "super::known_user::Column::UserId",
         on_update = "NoAction",
         on_delete = "NoAction"
     )]
-    User,
+    KnownUser,
 }
 
-impl Related<super::user::Entity> for Entity {
+impl Related<super::known_user::Entity> for Entity {
     fn to() -> RelationDef {
-        Relation::User.def()
+        Relation::KnownUser.def()
     }
 }
