+++ deepwell/src/models/forum_post_revision.rs
@@ -101,6 +101,14 @@ pub enum Relation {
     )]
     ForumThread1,
     #[sea_orm(
+        belongs_to = "super::known_user::Entity",
+        from = "Column::UserId",
+        to = "super::known_user::Column::UserId",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    KnownUser,
+    #[sea_orm(
         belongs_to = "super::site::Entity",
         from = "Column::SiteId",
         to = "super::site::Column::SiteId",
@@ -124,25 +132,17 @@ pub enum Relation {
         on_delete = "NoAction"
     )]
     Text1,
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::UserId",
-        to = "super::user::Column::UserId",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    User,
+}
+
+impl Related<super::known_user::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::KnownUser.def()
+    }
 }
 
 impl Related<super::site::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Site.def()
-    }
-}
-
-impl Related<super::user::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::User.def()
     }
 }
