+++ deepwell/src/models/message.rs
@@ -23,6 +23,14 @@ pub struct Model {
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
+        belongs_to = "super::known_user::Entity",
+        from = "Column::UserId",
+        to = "super::known_user::Column::UserId",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    KnownUser,
+    #[sea_orm(
         belongs_to = "super::message_record::Entity",
         from = "Column::RecordId",
         to = "super::message_record::Column::ExternalId",
@@ -32,14 +40,12 @@ pub enum Relation {
     MessageRecord,
     #[sea_orm(has_many = "super::message_report::Entity")]
     MessageReport,
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::UserId",
-        to = "super::user::Column::UserId",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    User,
+}
+
+impl Related<super::known_user::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::KnownUser.def()
+    }
 }
 
 impl Related<super::message_record::Entity> for Entity {
@@ -51,12 +57,6 @@ impl Related<super::message_record::Entity> for Entity {
 impl Related<super::message_report::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::MessageReport.def()
-    }
-}
-
-impl Related<super::user::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::User.def()
     }
 }
