+++ deepwell/src/models/message_recipient.rs
@@ -18,6 +18,14 @@ pub struct Model {
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
+        belongs_to = "super::known_user::Entity",
+        from = "Column::RecipientId",
+        to = "super::known_user::Column::UserId",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    KnownUser,
+    #[sea_orm(
         belongs_to = "super::message_record::Entity",
         from = "Column::RecordId",
         to = "super::message_record::Column::ExternalId",
@@ -25,25 +33,17 @@ pub enum Relation {
         on_delete = "NoAction"
     )]
     MessageRecord,
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::RecipientId",
-        to = "super::user::Column::UserId",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    User,
+}
+
+impl Related<super::known_user::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::KnownUser.def()
+    }
 }
 
 impl Related<super::message_record::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::MessageRecord.def()
-    }
-}
-
-impl Related<super::user::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::User.def()
     }
 }
