+++ deepwell/src/models/mod.rs
@@ -17,6 +17,7 @@ pub mod forum_post_lock;
 pub mod forum_post_revision;
 pub mod forum_thread;
 pub mod forum_thread_lock;
+pub mod known_user;
 pub mod message;
 pub mod message_draft;
 pub mod message_recipient;
@@ -41,3 +42,4 @@ pub mod text;
 pub mod text_block;
 pub mod user;
 pub mod user_role;
+pub mod wikidot_user;
