+++ deepwell/src/models/prelude.rs
@@ -15,6 +15,7 @@ pub use super::forum_post_lock::Entity as ForumPostLock;
 pub use super::forum_post_revision::Entity as ForumPostRevision;
 pub use super::forum_thread::Entity as ForumThread;
 pub use super::forum_thread_lock::Entity as ForumThreadLock;
+pub use super::known_user::Entity as KnownUser;
 pub use super::message::Entity as Message;
 pub use super::message_draft::Entity as MessageDraft;
 pub use super::message_recipient::Entity as MessageRecipient;
@@ -39,3 +40,4 @@ pub use super::text::Entity as Text;
 pub use super::text_block::Entity as TextBlock;
 pub use super::user::Entity as User;
 pub use super::user_role::Entity as UserRole;
+pub use super::wikidot_user::Entity as WikidotUser;
