+++ deepwell/src/models/user_role.rs
@@ -23,6 +23,22 @@ pub struct Model {
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
+        belongs_to = "super::known_user::Entity",
+        from = "Column::AssignedBy",
+        to = "super::known_user::Column::UserId",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    KnownUser2,
+    #[sea_orm(
+        belongs_to = "super::known_user::Entity",
+        from = "Column::UserId",
+        to = "super::known_user::Column::UserId",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    KnownUser1,
+    #[sea_orm(
         belongs_to = "super::role::Entity",
         from = "Column::RoleId",
         to = "super::role::Column::RoleId",
@@ -38,22 +54,6 @@ pub enum Relation {
         on_delete = "NoAction"
     )]
     Site,
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::AssignedBy",
-        to = "super::user::Column::UserId",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    User2,
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::UserId",
-        to = "super::user::Column::UserId",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    User1,
 }
 
 impl Related<super::role::Entity> for Entity {
