+++ deepwell/src/services/session/service.rs
@@ -31,6 +31,7 @@
 //! periodically.
 
 use super::prelude::*;
+use crate::models::known_user;
 use crate::models::session::{self, Entity as Session, Model as SessionModel};
 use crate::models::user::{self, Entity as User, Model as UserModel};
 use crate::utils::assert_is_csprng;
@@ -168,7 +169,7 @@ impl SessionService {
 
         let txn = ctx.transaction();
         let user_opt = User::find()
-            .join(JoinType::Join, user::Relation::Session.def())
+            .join(JoinType::Join, known_user::Relation::Session.def())
             .filter(
                 Condition::all()
                     .add(session::Column::SessionToken.eq(session_token))
